requirements.txt
numpy==1.23.1
pandas==1.4.3
pathlib==1.0.1
pyarrow==9.0.0
python-dateutil==2.8.2
pytz==2022.2
PyYAML==6.0
six==1.16.0
